wip: drop the executor namespace
