+++ src/kaleidoscope/simulator/reports/BootKeyboardReport.h
@@ -20,11 +20,9 @@
 
 #ifdef KALEIDOSCOPE_VIRTUAL_BUILD
 
-#include <stdint.h>
-
 #include "kaleidoscope/key_defs.h"
-#include "kaleidoscope/simulator/interface/reports/KeyboardReport_.h"
-#include "MultiReport/Keyboard.h"
+#include "kaleidoscope/simulator/interface/reports/BootKeyboardReport_.h"
+#include "BootKeyboard/BootKeyboard.h"
 
 // Undefine some macros defined by Arduino
 //
@@ -37,40 +35,39 @@
 
 namespace kaleidoscope {
 namespace simulator {
-namespace executor {
 
 class Simulator;
 
-/// @brief An interface hat facilitates analyzing keyboard reports.
+/// @brief An interface hat facilitates analyzing boot keyboard reports.
 ///
-class KeyboardReport : public interface::KeyboardReport_ {
+class BootKeyboardReport : public interface::BootKeyboardReport_ {
 
  public:
 
-  typedef HID_KeyboardReport_Data_t ReportDataType;
+  typedef HID_BootKeyboardReport_Data_t ReportDataType;
 
-  static constexpr uint8_t hid_report_type_ = HID_REPORTID_NKRO_KEYBOARD;
+  static constexpr uint8_t hid_report_type_ = HID_REPORTID_KEYBOARD;
 
   /// @brief Default consturctor.
   /// @details Creates an empty report.
   ///
-  KeyboardReport();
+  BootKeyboardReport();
 
   /// @brief Constructs based on a raw pointer to report data.
   /// @details Only use this if you know what you are doning!
   /// @param data The address where the report data starts.
   ///
-  KeyboardReport(const void *data);
+  BootKeyboardReport(const void *data);
 
   /// @brief Constructs based on a report data object.
   /// @param report_data The report data object to read.
   ///
-  KeyboardReport(const ReportDataType &report_data);
+  BootKeyboardReport(const ReportDataType &report_data);
 
   template<typename..._Args>
-  static std::shared_ptr<KeyboardReport> create(_Args &&... args) {
-    return std::shared_ptr<KeyboardReport> {
-      new KeyboardReport{std::forward<_Args>(args)...}
+  static std::shared_ptr<BootKeyboardReport> create(_Args &&... args) {
+    return std::shared_ptr<BootKeyboardReport> {
+      new BootKeyboardReport{std::forward<_Args>(args)...}
     };
   }
 
@@ -147,7 +144,6 @@ class KeyboardReport : public interface::KeyboardReport_ {
   ReportDataType report_data_;
 };
 
-} // namespace executor
 } // namespace simulator
 } // namespace kaleidoscope
